Update job status & error monitoring mechanism
All job filenames and ids in Airflow DB are generated
on the base of BioWardrobe uid. This allows to catch
the cases when new job file should be generated,
but some results of the previous run already exist
in Airflow DB or filesystem (as a job files).
System submits error to DB if:
 - duplicate dag_id is found
 - duplicate job file is found in 'new'/'running' folder
 - job file hasn't been run at all and was moved to 'fail'
Refactored to run all DB related functions from one script.
Some bugs are fixed

+++ basic_analysis/run_dna_func.py
@@ -2,8 +2,32 @@ import os
 import json
 import DefFunctions as util
 import collections
-from basic_analysis.exceptions import BiowFileNotFoundException, BiowJobException
-from basic_analysis.constants import LIBSTATUS, BOWTIE_INDICES, CHR_LENGTH
+from basic_analysis.exceptions import (BiowFileNotFoundException,
+                                       BiowJobException,
+                                       BiowWorkflowException)
+from basic_analysis.constants import (LIBSTATUS,
+                                      BOWTIE_INDICES,
+                                      CHR_LENGTH,
+                                      JOBS_NEW,
+                                      JOBS_RUNNING,
+                                      JOBS_FAIL)
+
+
+def check_job(db_settings, row, workflow, jobs_folder):
+    tasks, total = util.get_tasks(row[1], db_settings)
+    tasks = {k: v for k,v in tasks.iteritems() if v}
+    if not tasks:
+        failed_file = os.path.join(jobs_folder, JOBS_FAIL, os.path.splitext(os.path.basename(workflow))[0] + '-' + row[1] + '.json')
+        if os.path.isfile(failed_file):
+            raise BiowJobException(row[1], message="Failed to run job file. Check if correspondent workflow exists")
+        return (None, None)
+    if tasks.get("failed"):
+        raise BiowWorkflowException (row[1], message = tasks)
+    elif total > 0 and len(tasks.get("success", [])) == total:
+        return (LIBSTATUS["SUCCESS_PROCESS"], "Complete")
+    else:
+        return (LIBSTATUS["PROCESSING"], tasks)
+
 
 def get_control(db_settings, **kwargs):
     if not kwargs['control_id']: return None
@@ -43,8 +67,9 @@ def submit_job(db_settings, row, raw_data, indices, workflow, template_job, thre
     kwargs["chrom_length"] = os.path.join(kwargs["indices"], BOWTIE_INDICES, kwargs["genome"], CHR_LENGTH)
     kwargs["output_folder"] = os.path.join(kwargs["raw_data"], kwargs["uid"])
 
-    output_filename = os.path.join(jobs_folder, kwargs["workflow"] + '-' + kwargs["uid"] + '.json')
-
+    output_filename = os.path.join(jobs_folder, JOBS_NEW, kwargs["workflow"] + '-' + kwargs["uid"] + '.json')
+    running_filename = os.path.join(jobs_folder, JOBS_RUNNING, kwargs["workflow"] + '-' + kwargs["uid"] + '.json')
+    
     try:
         kwargs["control_file"] = get_control(db_settings, **kwargs)
     except BiowFileNotFoundException:
@@ -54,6 +79,10 @@ def submit_job(db_settings, row, raw_data, indices, workflow, template_job, thre
         raise BiowFileNotFoundException(kwargs["uid"])
     filled_job_object = util.remove_not_set_inputs(json.loads(template_job.format(**kwargs).replace("'True'",'true').replace("'False'",'false').replace('"True"','true').replace('"False"','false')))
     filled_job_str = json.dumps(collections.OrderedDict(sorted(filled_job_object.items())),indent=4)
+
+    # Check if file exists in job folder (running or new)
+    if os.path.isfile(output_filename) or os.path.isfile(running_filename):
+        raise BiowJobException(kwargs['uid'], message="Duplicate job file. Use ForceRun")
 
     try:
         with open(output_filename, 'w') as output_file:
